pyproject.toml
[project]
name = "partner-games-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (F), pycodestyle (E, W), isort (I)
select = ["F", "E", "W", "I"]

[tool.mypy]
python_version = "3.13"
strict = true
